requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
SQLAlchemy==2.0.39
pgvector==0.4.0
langchain==0.3.21
langchain-core==0.3.46
langchain-google-genai==2.1.0
langchain-postgres==0.0.13
langchain-text-splitters==0.3.7
langsmith==0.3.18
google-ai-generativelanguage==0.6.17
google-api-core==2.24.2
google-auth==2.38.0
numpy==2.2.4
pydantic==2.10.6
requests==2.32.3
python-dotenv==1.0.1
PyYAML==6.0.2
pytest==8.3.5
